+++ src/components/rental/image.svelte
@@ -1,0 +1,8 @@
+<script>
+	export let src;
+	export let alt;
+</script>
+
+<div class="image">
+	<img {src} {alt} />
+</div>